ab_example_contract_flipper::ffi

Module new

Structs

FlipperNewArgs

Data structure containing expected input for external method invocation, eventually calling flipper_new() on the other side by the host.

InternalArgs

Data structure containing expected input to flipper_new(), it is used internally by the contract, there should be no need to construct it explicitly except maybe in contract’s own tests

Constants

METADATA

Method metadata, see ContractMetadataKind for encoding details

Functions

flipper_new⚠

FFI interface into a method, called by the host.